requirements.txt
pyyaml
rich
markdown_it_py
mdurl
pygments
